tests: cover archive button via populated GoalTab story
Replace the standalone GoalBoardSections story with a full-tab
`CompleteWithBoard` story so the row Archive button (and its
RowActionButton siblings — Revive / Promote / Remove / Edit) render
in context alongside the goal header, budget tile, and de-emphasized
'Archive this goal' link.

GoalTab only mounts GoalBoardSections when `workspaceId` is set, and
the board reaches the renderer through `useGoalBoard → api.workspace
.getGoalBoard`. To populate that path in storybook:

- Add a `goalBoardSnapshots: Map<workspaceId, GoalBoardSnapshot>`
  option to `createMockORPCClient`; `getGoalBoard` returns the seeded
  snapshot for the calling workspace or falls back to empty entries.
- Wrap the new story in an `APIProvider` decorator backed by the mock
  client, with one upcoming / one completed / one archived entry.
- Tag the story with `CHROMATIC_SMOKE_MODES` so the restyle gains
  dual-theme baseline coverage in Chromatic.

`GoalTab.test.tsx` still mocks GoalBoardSections away, so this is the
only automated visual surface for the row-action buttons.
